remove isStream to the common lib.
