+++ lib/common.js
@@ -306,6 +306,13 @@ function formatQueryValue(key, value, options) {
   return [key, value];
 }
 
+function isStream(obj) {
+  return (typeof obj !== 'undefined') &&
+      (typeof a !== 'string') &&
+      (! Buffer.isBuffer(obj)) &&
+      _.isFunction(obj.setEncoding);
+}
+
 exports.normalizeRequestOptions = normalizeRequestOptions;
 exports.isBinaryBuffer = isBinaryBuffer;
 exports.mergeChunks = mergeChunks;
@@ -321,3 +328,5 @@ exports.percentEncode = percentEncode;
 exports.percentDecode = percentDecode;
 exports.matchStringOrRegexp = matchStringOrRegexp;
 exports.formatQueryValue = formatQueryValue;
+exports.isStream = isStream;
+
