+++ lib/delayed_body.js
@@ -14,12 +14,7 @@ var Transform = require('stream').Transform;
 var EventEmitter = require('events').EventEmitter;
 var noop = function () {};
 var util = require('util');
-var timers = require('timers');
-
-function isStream(obj) {
-  var is = obj && (typeof a !== 'string') && (! Buffer.isBuffer(obj)) && (typeof obj.setEncoding === 'function');
-  return is;
-}
+var common = require('./common');
 
 if (!Transform) {
   // for barebones compatibility for node < 0.10
@@ -56,7 +51,7 @@ function DelayedBody(ms, body) {
   var data = '';
   var ended = false;
 
-  if (isStream(body)) {
+  if (common.isStream(body)) {
     body.on('data', function (chunk) {
       data += Buffer.isBuffer(chunk) ? chunk.toString() : chunk;
     });
@@ -69,7 +64,7 @@ function DelayedBody(ms, body) {
   }
 
   setTimeout(function () {
-    if (isStream(body) && !ended) {
+    if (common.isStream(body) && !ended) {
       body.once('end', function () {
         self.end(data);
       });
