+++ lib/interceptor.js
@@ -99,7 +99,7 @@ Interceptor.prototype.reply = function reply(statusCode, body, headers) {
         if (body && typeof(body) !== 'string' &&
             typeof(body) !== 'function' &&
             !Buffer.isBuffer(body) &&
-            !isStream(body)) {
+            !common.isStream(body)) {
             try {
                 body = stringify(body);
                 if (!this.headers) {
@@ -496,10 +496,3 @@ Interceptor.prototype.socketDelay = function socketDelay(ms) {
     this.socketDelayInMs = ms;
     return this;
 };
-
-function isStream(obj) {
-    return (typeof obj !== 'undefined') &&
-        (typeof a !== 'string') &&
-        (! Buffer.isBuffer(obj)) &&
-        _.isFunction(obj.setEncoding);
-}
