+++ lib/request_overrider.js
@@ -44,11 +44,6 @@ function setHeader(request, name, value) {
   }
 }
 
-function isStream(obj) {
-  var is = obj && (typeof obj !== 'string') && (!Buffer.isBuffer(obj)) && (typeof obj.setEncoding === 'function');
-  return is;
-}
-
 //  Sets request headers of the given request. This is needed during both matching phase
 //  (in case header filters were specified) and mocking phase (to correctly pass mocked
 //  request headers).
@@ -317,7 +312,7 @@ function RequestOverrider(req, options, interceptors, remove, cb) {
       //  of response buffers which should be mocked one by one.
       //  (otherwise decompressions after the first one fails as unzip expects to receive
       //  buffer by buffer and not one single merged buffer)
-      if(common.isContentEncoded(response.headers) && ! isStream(interceptor.body)) {
+      if(common.isContentEncoded(response.headers) && ! common.isStream(interceptor.body)) {
 
         if (interceptor.delayInMs) {
           emitError(new Error('Response delay is currently not supported with content-encoded responses.'));
@@ -400,7 +395,7 @@ function RequestOverrider(req, options, interceptors, remove, cb) {
           responseBody = new DelayedBody(interceptor.delayInMs, responseBody);
         }
 
-        if (isStream(responseBody)) {
+        if (common.isStream(responseBody)) {
           debug('response body is a stream');
           responseBody.pause();
           responseBody.on('data', function(d) {
@@ -492,7 +487,7 @@ function RequestOverrider(req, options, interceptors, remove, cb) {
             req.emit('response', response);
           }
 
-          if (isStream(responseBody)) {
+          if (common.isStream(responseBody)) {
             debug('resuming response stream');
             responseBody.resume();
           }
